backend/requirements/base.txt
django==5.2.12
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
drf-spectacular==0.29.0
